backend/requirements_no_audioop.txt
aiofiles==24.1.0
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
bcrypt==4.3.0
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
dnspython==2.8.0
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.115.12
fonttools==4.58.4
h11==0.16.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
motor==3.7.1
numpy==2.3.1
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
passlib==1.7.4
pillow==11.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pymongo==4.14.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
qrcode==8.2
redis==6.4.0
reportlab==4.4.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3

# --- Django stack (faltaba en tu lista original) ---
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0

# --- MFA / OTP (faltaba: pyotp) ---
pyotp==2.9.0
